refactor into util package
