+++ otpsecure/src/main/java/com/ecertic/otpsecure/utils/OtpSecureUtils.java
@@ -1,4 +1,4 @@
-package com.ecertic.otpsecure;
+package com.ecertic.otpsecure.utils;
 
 import android.text.TextUtils;
