+++ otpsecure/src/test/java/com/ecertic/otpsecure/OtpSecureApiHandlerTest.java
@@ -125,5 +125,4 @@ public class OtpSecureApiHandlerTest {
         }
     }
 
-
 }
